requirements.txt
streamlit>=1.28.0
psycopg2-binary>=2.9.9
pandas>=2.0.0
plotly>=5.18.0
bcrypt>=4.1.0
python-dotenv>=1.0.0
